+++ src/server/subscribe-callbacks.js
@@ -1,6 +1,8 @@
 var bindAll = require('101/bind-all')
 var debug = require('debug')('primus-graphql:subscribe-callbacks')
 
+var QueryExecutor = require('./query-executor.js')
+
 module.exports = SubscribeCallbacks
 
 function SubscribeCallbacks (payload, queryExecutor, responder) {
@@ -32,8 +34,12 @@ SubscribeCallbacks.prototype.onNext = function (next) {
   var self = this
   var id = this._payload.id
   debug('onNext:', id, next)
-  var rootValue = next
-  return this._queryExecutor.execute(this._payload, rootValue).then(function (payload) {
-    self._responder.sendEvent(id, 'next', payload.data)
+  var parsed = QueryExecutor.parseQuery(self._payload.query)
+  var subscriptionFieldName = parsed.definitions[0].selectionSet.selections[0].name.value
+  var rootValue = {}
+  rootValue[subscriptionFieldName] = next
+  debug('onNext rootValue:', id, rootValue)
+  return this._queryExecutor.execute(this._payload, rootValue).then(function (result) {
+    self._responder.sendEvent(id, 'next', result.data)
   }).catch(this.onError)
 }