+++ test-browser/karmaconf.js
@@ -60,7 +60,7 @@ var json = {
   colors: true,
 
   // enable / disable watching file and executing tests whenever any file changes
-  autoWatch: false,
+  autoWatch: process.env.WATCH,
 
   // start these browsers
   // available browser launchers: https://npmjs.org/browse/keyword/karma-launcher
@@ -68,7 +68,7 @@ var json = {
 
   // Continuous Integration mode
   // if true, Karma captures browsers, runs the tests and exits
-  singleRun: true,
+  singleRun: !process.env.WATCH,
 
   // Concurrency level
   // how many browser should be started simultaneous